+++ applovin_max/android/src/main/java/com/applovin/applovin_max/AppLovinMAXAdViewFactory.java
@@ -45,12 +45,13 @@ public class AppLovinMAXAdViewFactory
         // Optional params
         Integer ad_view_id = params.containsKey( "ad_view_id" ) ? (Integer) params.get( "ad_view_id" ) : null;
         int adViewId = ad_view_id != null ? ad_view_id : 0;
+        boolean isAdaptiveBannerEnabled = Boolean.TRUE.equals( params.get( "is_adaptive_banner_enabled" ) ); // Defaults to true
         boolean isAutoRefreshEnabled = Boolean.TRUE.equals( params.get( "is_auto_refresh_enabled" ) ); // Defaults to true
         String placement = params.containsKey( "placement" ) ? (String) params.get( "placement" ) : null;
         String customData = params.containsKey( "custom_data" ) ? (String) params.get( "custom_data" ) : null;
         Map<String, Object> extraParameters = params.containsKey( "extra_parameters" ) ? (Map<String, Object>) params.get( "extra_parameters" ) : null;
         Map<String, Object> localExtraParameters = params.containsKey( "local_extra_parameters" ) ? (Map<String, Object>) params.get( "local_extra_parameters" ) : null;
 
-        return new AppLovinMAXAdView( viewId, adUnitId, adViewId, adFormat, isAutoRefreshEnabled, placement, customData, extraParameters, localExtraParameters, messenger, sdk, context );
+        return new AppLovinMAXAdView( viewId, adUnitId, adViewId, adFormat, isAdaptiveBannerEnabled, isAutoRefreshEnabled, placement, customData, extraParameters, localExtraParameters, messenger, sdk, context );
     }
 }