+++ applovin_max/ios/Classes/AppLovinMAXAdViewWidget.h
@@ -12,6 +12,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)setCustomData:(nullable NSString *)customData;
 - (void)setExtraParameters:(nullable NSDictionary<NSString *, id> *)extraParameters;
 - (void)setLocalExtraParameters:(nullable NSDictionary<NSString *, id> *)localExtraParameters;
+- (void)setAdaptiveBannerEnabled:(BOOL)adaptiveBannerEnabled;
 - (void)setAutoRefreshEnabled:(BOOL)autoRefreshEnabled;
 
 - (void)loadAd;