+++ applovin_max/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Versions
 
+## x.x.x
+* Add platform implementation for adaptive banner.
 ## 4.4.0
 * Depends on Android SDK v13.2.0 and iOS SDK v13.2.0.
 ## 4.3.1